backend/requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
flasgger
python-dotenv
psycopg2-binary
alembic
